test(igxgrid): test for selecting conditions and input is empty #1118
